add check

+++ src/main/java/eu/fbk/dslab/playful/engine/rest/LearningModuleController.java
@@ -1,5 +1,6 @@
 package eu.fbk.dslab.playful.engine.rest;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,23 +17,47 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import eu.fbk.dslab.playful.engine.exception.UnauthorizedException;
+import eu.fbk.dslab.playful.engine.manager.EntityManager;
 import eu.fbk.dslab.playful.engine.model.LearningModule;
+import eu.fbk.dslab.playful.engine.model.LearningScenario;
 import eu.fbk.dslab.playful.engine.repository.LearningModuleRepository;
+import eu.fbk.dslab.playful.engine.repository.LearningScenarioRepository;
+import eu.fbk.dslab.playful.engine.security.SecurityHelper;
+import eu.fbk.dslab.playful.engine.security.UserRole.Role;
 
 @RestController
 public class LearningModuleController {
 	@Autowired
 	LearningModuleRepository learningModuleRepository;
+	@Autowired
+	LearningScenarioRepository learningScenarioRepository;
+	
+	@Autowired
+	EntityManager entityManager;
+	
+	@Autowired
+	SecurityHelper securityHelper;
 	
 	@GetMapping("/api/modules")
 	public Page<LearningModule> getList(
 			@RequestParam(required = false) List<String> ids,
-			@RequestParam(required = false) String learningScenarioId) {
-		List<LearningModule> list = null;
+			@RequestParam(required = false) String learningScenarioId) throws Exception {
+		List<LearningModule> list = new ArrayList<>();
 		if(ids != null) {
 			list = learningModuleRepository.findByIdIn(ids);
 		} else if(learningScenarioId != null) {
-			list = learningModuleRepository.findByLearningScenarioId(learningScenarioId, Sort.by(Direction.ASC, "position"));
+			LearningScenario ls = learningScenarioRepository.findById(learningScenarioId).orElse(null);
+			if(ls != null) {
+				securityHelper.checkRole(ls.getDomainId(), Role.domain, Role.educator);
+				if(securityHelper.hasRole(ls.getDomainId(), Role.domain)) {
+					list = learningModuleRepository.findByLearningScenarioId(learningScenarioId, Sort.by(Direction.ASC, "position"));
+				} else {
+					if(entityManager.checkEducator(learningScenarioId)) {
+						list = learningModuleRepository.findByLearningScenarioId(learningScenarioId, Sort.by(Direction.ASC, "position"));	
+					}
+				}
+			}
 		} else {
 			list = learningModuleRepository.findAll();
 		}
@@ -40,27 +65,47 @@ public class LearningModuleController {
 	}
 	
 	@GetMapping("/api/modules/{id}")
-	public LearningModule getOne(@PathVariable String id) {
-		return learningModuleRepository.findById(id).orElse(null);
+	public LearningModule getOne(@PathVariable String id) throws Exception {
+		LearningModule entity = learningModuleRepository.findById(id).orElse(null);
+		if(entity != null) {
+			securityHelper.checkRole(entity.getDomainId(), Role.domain, Role.educator);
+			if(securityHelper.hasRole(entity.getDomainId(), Role.educator)) {
+				if(!entityManager.checkEducator(entity.getLearningScenarioId())) {
+					throw new UnauthorizedException("role not found");
+				}
+			}
+		}
+		return entity;
 	}
 	
 	@PostMapping("/api/modules")
-	public LearningModule create(@RequestBody LearningModule lLearningModule) {
-		return learningModuleRepository.save(lLearningModule);
+	public LearningModule create(@RequestBody LearningModule learningModule) throws Exception {
+		if(securityHelper.hasRole(learningModule.getDomainId(), Role.domain) || 
+				entityManager.checkEducator(learningModule.getLearningScenarioId())) {
+			return learningModuleRepository.save(learningModule);
+		}
+		throw new UnauthorizedException("role not found");
 	}
 	
 	@PutMapping("/api/modules/{id}")
 	public LearningModule update(@PathVariable String id,
-			@RequestBody LearningModule lLearningModule) {
-		lLearningModule.setId(id);
-		return learningModuleRepository.save(lLearningModule);
+			@RequestBody LearningModule learningModule) throws Exception {
+		learningModule.setId(id);
+		if(securityHelper.hasRole(learningModule.getDomainId(), Role.domain) || 
+				entityManager.checkEducator(learningModule.getLearningScenarioId())) {
+			return learningModuleRepository.save(learningModule);
+		}
+		throw new UnauthorizedException("role not found");
 	}
 	
 	@DeleteMapping("/api/modules/{id}")
-	public LearningModule delete(@PathVariable String id) {
+	public LearningModule delete(@PathVariable String id) throws Exception {
 		LearningModule learningModule = learningModuleRepository.findById(id).orElse(null);
 		if(learningModule != null) {
-			learningModuleRepository.deleteById(id);
+			if(securityHelper.hasRole(learningModule.getDomainId(), Role.domain) || 
+					entityManager.checkEducator(learningModule.getLearningScenarioId())) {
+				learningModuleRepository.deleteById(id);
+			}
 		}
 		return learningModule;
 	}
